[ci] delete JOB_BASE_NAME
`JOB_BASE_NAME` was a holdover from jenkins compatibility. Eventually,
it morphed to be always set to the build enviroment + `-test` or
`-build`, and we used it to detect whether we were in a build or test.

That's sort of pointless, so removing and fixing up the few remaining
use cases.

ghstack-source-id: 69285834e30326e7ab364373413234a852e8165b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80046